+++ resources/help/history.md
@@ -9,5 +9,18 @@ history log.  lliascript may also add remarks to the history, particularly
 in the case of an error.
 
 
-The history log is executable Python and may be saved, edited and reloaded
-to recreate the current state of Llia.  
+The history log is executable Python and may be saved and reloaded to
+recreate a Llia state.
+
+
+If A GUI is enabled a basic text editor is available under the
+File/lliascript menu.
+
+
+The editor tools include:
+
+-   Open - Read a python file into the editor
+-   Save - Save the current editor contents to a file
+-   Exec - Execute the current editor contents as a lliascript batch file.
+-   Compose - Creates a python script which may be used to recreate the
+              current state of Llia.